Work around parser bugs around very large negative numbers
Fixes #1574

+++ src/Mutator/Number/DecrementInteger.php
@@ -86,7 +86,10 @@ DIFF
         $value = $node->value - 1;
 
         if ($parentNode instanceof Node\Expr\UnaryMinus) {
-            $value = $node->value + 1;
+            // PHP Parser reads negative number as a pair of minus sign and a positive int,
+            // but positive part of PHP_INT_MIN leads to an overflow into float. To work
+            // around this we have to cast the result value back to int.
+            $value = (int) ($node->value + 1);
         }
 
         yield new Node\Scalar\LNumber($value);
